fix: don't report Unreachable errors
Though this would be nice to have, it's not worth the effort of keeping
track of when a statement is actually unreachable after a return
statement, since in some cases it can be somewhat complex. For example,

```go
if cond {
	return
}

something // should not be unreachable (depends on cond)
```

and the same would also arguably be true for a for-loop, but:

```go
{
	return
}

something // this is unreachable
```

Knowing this, it's easier to just avoid false positives altogether by
not tracking this error at all and never reporting anything of this
kind. The Go compiler can catch it.