feat: use new defer branch backtrace model
Instead of tracking a single "returning" state in the global context,
allow deferring the popping of the current backtrace until a later
stage. For example, normally the branch backtrace corresponding to an if
condition expires at the end of the if block, but if there's a
flow-altering statement inside (like a return statement) then the
backtrace must be kept until the end of the surrounding function.