feat(taint): semi-fine-grained array analysis
This adds support for remembering (up to one level) backtraces
associated with each index of an array, as long as they are accessed
with a "constant" index (in practice, basic compositions of integer
literals). The symbol's normal backtrace is also kept for when this is
not true, e.g. `arr[x] = secret`.